+++ src/main/java/com/eugene/crude/crude/practic/controller/ControllerImpl/UserControllerImpl.java
@@ -4,8 +4,8 @@ package com.eugene.crude.crude.practic.controller.ControllerImpl;
 import com.eugene.crude.crude.practic.controller.UserController;
 import com.eugene.crude.crude.practic.model.User;
 import com.eugene.crude.crude.practic.repository.RegionRepository;
-import com.eugene.crude.crude.practic.repository.repositoryIO.RegionRepositoryImpl;
-import com.eugene.crude.crude.practic.repository.repositoryIO.UserRepositoryImpl;
+import com.eugene.crude.crude.practic.repository.bdRepositoty.RegionRepositoryImpl;
+import com.eugene.crude.crude.practic.repository.bdRepositoty.UserRepositoryImpl;
 
 import java.io.IOException;
 import java.util.List;
@@ -24,13 +24,13 @@ RegionRepository regionRepository = new RegionRepositoryImpl();
 
     public void deleteById(String str) throws IOException {
 
-        userRepository.deleteById(Long.parseLong(str));
+        userRepository.deleteById(Integer.parseInt(str));
     }
 
 
     public User getElementById(String str) throws IOException {
 
-       User user = userRepository.getById(Long.parseLong(str));
+       User user = userRepository.getById(Integer.parseInt(str));
         if (user == null)
             return null;
         else {
